pyproject.toml
[project]
name = "company-ai"
version = "0.1.0"
description = "사내 AI 어시스턴트 플랫폼"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.9",
    # Agent
    "langgraph>=0.5.1",
    "langchain>=0.3.26",
    "langchain-text-splitters>=0.3.8",
    # LLM
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    # Vector DB
    "qdrant-client>=1.14.3",
    "fastembed>=0.7.1",
    # Document readers
    "notion-client>=2.2.1",
    "atlassian-python-api>=3.41.0",
    "msal>=1.28.0",
    "pymupdf>=1.23.14",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.13.4",
    # Utils
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "httpx>=0.27.0",
    "tqdm>=4.65.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
